pyproject.toml
[build-system]
requires = [ "poetry-core>=1",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "autodisc-server-libs"
version = "0.1.0"
description = "Libraries needed to run the autodisc-server computational container."
authors = [ "LIN Jesse <jesse.lin@inria.fr>",]
readme = "README.md"
[[tool.poetry.packages]]
include = "auto_disc"

[[tool.poetry.packages]]
include = "auto_disc_db"

[[tool.poetry.packages]]
include = "adtool_default"

[tool.poetry.dependencies]
addict = "2.4.0"
filetype = "1.2.0"
graphviz = "0.20.1"
imageio = "2.34.0"
imageio-ffmpeg = "0.4.9"
matplotlib = "^3.8"
mergedeep = "1.3.4"
neat-python = "0.92"
numpy = "^1.26"
pexpect = "^4.8.0"
pillow = "10.2.0"
python = ">=3.9,<3.13"
requests = "^2"
sqlalchemy = "^2.0"
tinydb = "4.8.0"
toml = "^0.10.2"
torch = "^1.7.1"
urllib3 = "2.2.1"
watchdog = "^4.0.0"
Werkzeug= "^2.3"

[tool.poetry.group.test.dependencies]
coverage = "^7.2.7"
pytest = "^7.3.1"
pytest-mock = "^3.10.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
debugpy = "^1.6.7"
isort = "^5.12.0"

[tool.poetry.group.docs.dependencies]
pdoc = "^13.1.1"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
